refactor: merge touch and face id enrolment due to duplicate code
